fix(tests): separate singleton tests
